==> 01 Scripts/01 - IC Data Import & Descriptive Analysis.ipynb <==
# TOC
1. Importing libraries   
2. Importing data   
3. Examining the data through Python   
4. Removing a column   
5. Answering question 11   

## Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## Importing Data

df = pd.read_csv(r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis\02 Data\Original\orders.csv', index_col = False)

path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

path

## Examining Data

# Checking imported orders data - first five
df.head()

# Checking imported orders data - last five
df.tail()

# Checking how many rows and columns there are in the dataframe
df.shape

# Pulling initial information about the dataframe
df.info

# Getting the data types of all columns
df.info()

# Pulling descriptive statistics about each column
df.describe()

## Removing a column

# Choosing columns to keep - dropping eval_set
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'orders.csv'), usecols = vars_list)

# Checking column is gone
df.head()

df.info(5)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'products.csv'), index_col = False)

# Checking first 20 rows of df_prods
df_prods.head(20)

# Checking last 35 rows of df_prods
df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()

df_prods.dtypes

## Question 11 Answers

aisle_id maximum value is 134

data type of department_id column is int64, or a 64 bit integer.
