==> 01 Scripts/02 - Wrangling and Subsetting.ipynb <==
# TOC
1. Importing libraries   
2. Importing data   
3. Transposing department data   
4. Setting new header   
5. Creating data dictionary   
6. Pulling 'snacks' subset   
7. Finding the busiest hour for Instacart sales   
8. Creating a subset for breakfast items   
9. Creating a subset for dinner party items   
10. Pulling User ID 1 data and analyzing it   
11. Exporting data

## Wrangling Exercise

# Importing libraries

import pandas as pd
import numpy as np
import os

# Defining data location path
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

# Checking path
path

# Importing departments.csv dataset
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'departments.csv'), index_col = False)

df_dep.head()

# Importing products.csv dataset
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'products.csv'), index_col = False)

# Setting vars_list for orders.csv
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Importing orders.csv dataset
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'orders.csv'), usecols = vars_list)

# Transpose df_dep
df_dep.T

df_dep_t =  df_dep.T

df_dep_t

# Resetting index
df_dep_t.reset_index()

# Setting new header
new_header =  df_dep_t.iloc[0]

# New dataframe from row 1 onward
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Inserting new header
df_dep_t_new.columns = new_header

df_dep_t_new

# Creating data dictionary
data_dict =  df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

# Pulling subset 'Snacks'
df_snacks = df_prods[df_prods['department_id']==19]

df_snacks

## Task Steps

# Examining df_ords
df_ords.head()

# Setting columns that don't need to be analyzed numerically
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['user_id'] = df_ords['user_id'].astype('str')

# Renaming non-intuitive name
df_ords.rename(columns={'order_dow' : 'order_day_of_week'}, inplace = True)

df_ords.head()

# Finding busiest hour
data = df_ords['order_hour_of_day']

data.value_counts()

The busiest hour for orders is 10 AM

# Determining value of department ID 4
print(data_dict.get('4'))

# Breakfast item subset creation
data_dict

df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast.head()

This isn't really going to tell you anything about SALES, just about the products, but looking at df_ords and what headers are there, I'm not seeing a way to tie the two together yet so you can get information about orders that contained breakfast items. 

# Dinner party subset creation
data_dict

df_dinnerparty = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinnerparty.head()

# Total count of df_dinnerparty
df_dinnerparty.shape

df_dinnerparty has 7,650 rows. 

# User ID 1 information generation
df_userone = df_ords[df_ords['user_id']=='1']

df_userone

# Basic stats about user_id 1
df_userone.describe()

user_id 1 has placed 11 orders. They usually place orders at the beginning of the week. They tend to place their orders in the morning. Their average length between orders is 19 days. 

# Exporting orders dataframe
df_ords.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'orders_wrangled.csv'))

# Exporting departments dataframe
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'departments_wrangled.csv'))

df_ords
