==> 01 Scripts/03 - Data Consistency.ipynb <==
# TOC
1. Preparation & Data import
2. Exercise
3. View descriptive stats
4. Removing unnecessary column
5. Checking for mixed type columns
6. Finding missing values and analyzing them
7. Checking for duplicates
8. Exporting dataframes

### Prep

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

path

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Cleaned', 'old', 'orders_wrangled.csv'), index_col = False)

### Exercise

# Create a dataframe
df_test =  pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Finding missing values
df_prods.isnull().sum()

# Isolating null values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Creating df_prods with no null values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods.shape

df_prods_clean.shape

# Searching for duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean_no_dupes = df_prods_clean.drop_duplicates()

df_prods_clean_no_dupes.shape

df_prods_clean_no_dupes.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'products_checked.csv'))

### Task

#### Step 2

# Viewing descriptive statistics for df_ords

df_ords.describe()

# Re-viewing df_ords due to new column appearing
df_ords

# Dropping generated repeated column Unnamed: 0 and re-importing
vars_list = ['order_id', 'user_id', 'order_number', 'order_day_of_week', 'order_hour_of_day', 'days_since_prior_order']

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Cleaned', 'old', 'orders_wrangled.csv'), usecols = vars_list)

df_ords

# Setting column types to exclude from descriptive statistics
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords.describe()

# Checking that zero is an applicable number for order_hour_of_day (ie, in place of midnight)
for col in df_ords:
  print(df_ords['order_hour_of_day'].unique())

I may just be reading the numbers incorrectly, but the max on days_since_prior_order only showing a 3 seems to be very very low. All of those numbers seem to be low, given that we looked at the days_since_prior_order for user_id 1 and they had double digit days. 

#### Step 3

# Checking for mixed-types columns
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#### Step 5

# Checking for missing values within df_ords
df_ords.isnull().sum()

days_since_prior_order has NaN for a number of values. My theory is that this is what generates on the first order for each customer. 

#### Step 6

# Testing my theory about missing values in days_since_prior_order
df_ords['user_id'].nunique()

REVISED: Because in this case the null value works as a flag to show that this is the customer's first order, it serves a specific purpose and should not be revised out. Thus, the "missing values" in this case are not really "missing" values and do not need to be further addressed. 

#### Step 7

# Pulling duplicates into new dataframe
df_ords_dupes = df_ords[df_ords.duplicated()]

# Checking new duplicates dataframe
df_ords_dupes.shape

#### Step 8

Pulling the duplicates out of df_ords yielded no full duplicates

#### Step 9

# Exporting final cleaned data
df_ords.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'orders_cleaned.csv'))

df_prods_clean_no_dupes.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'products_cleaned.csv'))
