==> 01 Scripts/04 - Combining & Exporting Data (Prelim).ipynb <==
# TOC
1. Prep & Import
2. Verify dimensions of dataset
3. Merge datasets

### Prep

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Cleaned', 'orders_cleaned.csv'))

df_ords

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Cleaned', 'products_cleaned.csv'))

df_prods

### Exercise

# Verify dimensions of df_ords
df_ords.shape

df_ords.head()

# Import orders_products_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'orders_products_prior.csv'), index_col = False)

# Verify dimensions of df_ords_prior
df_ords_prior.shape

df_ords_prior.head()

# Merging df_ords & df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Verifying inner join
df_merged_large['_merge'].value_counts()

df_merged_large.head()

df_merged_large.drop(columns='_merge')

df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_combined.pkl'))
