==> 01 Scripts/05 - Combining & Exporting Data (Main).ipynb <==
# TOC
1. Prep & Import   
2. Merge datasets   
3. Check new dataset   
4. Export new merged dataset as pickle   

### Prep

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Cleaned', 'products_cleaned.csv'))

df_prods

#### 4.6.3 - Importing Pickle

df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_combined.pkl'))

#### 4.6.4 - Checking Imported Dataframe

df_ords_prods.shape

#### 4.6.5 - Merging df_prods & df_ords_prods

# Merging dataframes on product_id column
orders_products_merged = df_ords_prods.merge(df_prods, on='product_id')

orders_products_merged.head()

# Checking merge rates - inner
orders_products_merged['_merge'].value_counts()

orders_products_merged_outer = df_ords_prods.merge(df_prods, on='product_id', how='outer')

# Checking merge rates - outer
orders_products_merged_outer['_merge'].value_counts()

#### 4.6.7 - Exporting new dataframe

# Removing _merge flag for tidiness
orders_products_merged.drop(columns='_merge')

# Exporting merged dataframe as pickle due to size
orders_products_merged.to_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_merged.pkl'))
