==> 01 Scripts/06 - Deriving New Variables.ipynb <==
# TOC
1. Prep & Import   
2. Create price buckets   
3. Create busiest days flag   
4. Create busiest hours flag   
5. Export pickle   

## Prep

# Importing libraries
import pandas as pd
import numpy as nm
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_merged.pkl'))

df_ords_prods.head()

## Exercise

#### Setting up price buckets

# defining df for if-then
df = df_ords_prods[:1000000]

# Creating price range function
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High range'
    else: return 'Not enough data'

# Applying the price range function
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

# Running price range function on whole dataframe via loc()
df_ords_prods.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods['price_range_loc'].value_counts(dropna = False)

#Creating busiest_day for-loop
df_ords_prods['order_day_of_week'].value_counts(dropna = False)

result = []

for value in df_ords_prods['order_day_of_week']:
    if value == 0:
        result.append("Busiest day")
    elif value == 4:
        result.append("Least busy")
    else:
        result.append("Regularly busy")

result

df_ords_prods['busiest_day'] = result

df_ords_prods.head()

df_ords_prods['busiest_day'].value_counts(dropna = False)

## Task

#### 4.7.2 - Creating 'Busiest Days' Column

# Updating busiest_day column to busiest days
df_ords_prods['order_day_of_week'].value_counts(dropna = False)

result = []

for value in df_ords_prods['order_day_of_week']:
    if value == 0 or value == 1:
        result.append("Busiest day")
    elif value == 4 or value == 3:
        result.append("Least busy")
    else:
        result.append("Regularly busy")

df_ords_prods['busiest_days'] = result

#### 4.7.3 - Checking 'Busiest Days' for Consistency

df_ords_prods['busiest_days'].value_counts(dropna = False)

Verifying here that the regularly busy label has fewer occurrences, and busiest/least busiest both have more (ROUGHLY twice as many, since we're adding another value to both). 

df_ords_prods.head()

#### 4.7.4 - Identifying the busiest hours of the day

# Finding busiest times of day
df_ords_prods['order_hour_of_day'].value_counts(dropna = False)

# Aggregating busiest period data
resulthr = []

for value in df_ords_prods['order_hour_of_day']:
    if value == 10 or value == 11 or value == 14:
        resulthr.append("Most orders")
    elif value == 3 or value == 4 or value == 2:
        resulthr.append("Fewest orders")
    else:
        resulthr.append("Average orders")

# Creating column for busiest hours
df_ords_prods['busiest_period_of_day'] = resulthr

#### 4.7.5 - Frequency for Busiest Period of Day column

# Checking frequency for busiest_period_of_day
df_ords_prods['busiest_period_of_day'].value_counts(dropna = False)

df_ords_prods.head()

# Removing extraneous columns
df_ords_prods.drop(columns=['_merge', 'busiest_day'])

# Saving pickle
df_ords_prods.to_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_busiest.pkl'))

df_ords_prods['price_range_loc'].value_counts(dropna = False)
