==> 01 Scripts/07 - Grouping & Aggregating Data (Prelim).ipynb <==
# TOC
1. Prep & import  
2. Create a subset of the data   
3. Create the max_order column   
4. Create customer loyalty flag   

## Prep

# Importing libraries
import pandas as pd
import numpy as nm
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Cleaned', 'orders_products_busiest.pkl'))

## Exercise

# Creating subset
df = ords_prods_merge[:1000000]

df.shape

df.head(10)

df.groupby('product_name')

df.groupby('department_id').agg({'order_number':['mean']})

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(nm.max)

ords_prods_merge.head(15)

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id','loyalty_flag','order_number']].head(60)

ords_prods_merge.head()

ords_prods_merge.drop(columns=['_merge','busiest_day'])

ords_prods_merge.head(10)
