==> 01 Scripts/08 - Grouping & Aggregating (Main).ipynb <==
# TOC
1. Prep and Import
2. Create loyalty marker  
3. Find average orders per department  
4. Create average price column  
5. Set low & high spender flags   
6. Create frequency flag   
7. Export pickle   

## Prep

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing data
path = path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Cleaned', 'ords_prods_merge_grouped.pkl'))

ords_prods_merge.head(15)

ords_prods_merge.shape

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

## Task

#### 4.8.2 - Finding average orders for each department for full dataframe

# Calculating average orders for each department
ords_prods_merge.groupby('department_id').order_number.mean()

#### 4.8.3 - Observations of dataframe vs subset

The entire dataframe contains ALL departments, which also changes which department has the lowest number of orders. In the subset, it was department 17, and for the full dataframe, it's department 5.

It also changes the highest-selling department - in the subset it was department 16, here it's department 21. 

#### 4.8.5 - Investigating spending habits of customer types

# Pulling basic stats for each loyalty category
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

I would have expected the opposite conclusion to happen here, but looking at the results and thinking about use cases - new customers may be purchasing higher dollar items that they need in an 'emergency' situation, whereas the loyal customers are likely purchasing across more departments & categories, which is going to even out their average product cost. 

#### 4.8.6 - Creating a spending flag

# Creating average price column
ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# Checking new column
ords_prods_merge.head(20)

# Setting low spender flag
ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spender_type'] = 'Low spender'

# Setting high spender flag
ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spender_type'] = 'High spender'

# Verification
ords_prods_merge.shape

ords_prods_merge['spender_type'].value_counts(dropna = False)

Total of high spender and low spender is 32,404,859 - matching shape of dataframe. 

The question here then becomes HOW much higher the high spenders are (ex: sure, the Kia Rio is a very popular car but how much more profit is a car company making every time they sell a ... BMW? I drive an '05 Civic, the car analogy wasn't my best choice), and is it a better use of dollars to get the high spenders to spend MORE or try to coax more out of the low spenders? 

#### 4.8.7 - Frequent vs Non-Frequent Customers

# Creating order regularity column
ords_prods_merge['frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# Checking new column
ords_prods_merge.head(20)

# Setting non-frequent buyer flag
ords_prods_merge.loc[ords_prods_merge['frequency'] > 20, 'freq_flag'] = 'Non-frequent customer'

# Setting regular buyer flag
ords_prods_merge.loc[ords_prods_merge['frequency'] <= 20, 'freq_flag'] = 'Regular customer'

# Setting frequent buyer flag
ords_prods_merge.loc[ords_prods_merge['frequency'] <= 10, 'freq_flag'] = 'Frequent customer'

# Verification
ords_prods_merge.shape

ords_prods_merge['freq_flag'].value_counts(dropna = False)

#### 4.8.9 - Exporting Data

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Cleaned', 'ords_prods_merge_mktg.pkl'))
