==> 01 Scripts/09 - Final Report Data Prep.ipynb <==
# TOC
1. Prep & import   
2. Wrangle customer data   
3. Isolate null values   
4. Drop customer name fields as they're not necessary for analysis   
5. Import orders/products data   
6. Merge with customer data   
7. Save fully merged data  

## Prep

# Importing libraries
import pandas as pd
import numpy as np
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Instacart Basket Analysis'

insta_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original', 'customers.csv'))

# Verifying
insta_cust.shape

## Wrangling

# General information
insta_cust.head()

# Checking data types
insta_cust.dtypes

insta_cust.info()

# Cleaning up columns
insta_cust['customer_name'] = insta_cust['First Name'] + ' ' + insta_cust['Surnam']

insta_cust.head()

# Moving customer name column to more logical position
insta_cust = insta_cust[['user_id','customer_name','Gender','STATE','Age','date_joined','n_dependants','fam_status','income']]

insta_cust.head()

insta_cust.rename(columns={'Gender':'gender', 'STATE':'state', 'Age':'age'},inplace = True)

insta_cust.head()

insta_cust.rename(columns={'n_dependants':'family_members','fam_status':'marital_status'},inplace = True)

insta_cust.head()

# Checking for duplicates
duplicate = insta_cust[insta_cust.duplicated()]

duplicate

# Investigating descriptive statistics
insta_cust.describe()

# Checking null
insta_cust.isnull().sum()

# Isolating null values
insta_null = insta_cust[insta_cust['customer_name'].isnull()==True]

insta_null

As "customer name" is the only column missing data, and the customer name can be pulled from other data (for instance, for a loyalty program), that column can be safely dropped as it is not necessary to analysis the way that other information included here could be (which gender uses instacart more, where are instacart's biggest and smallest state sales, are membership numbers going up or down by year, etc). 

# Dropping customer name column
insta_cust = insta_cust.drop(columns='customer_name')

# Saving CSV
insta_cust.to_csv(os.path.join(path, '02 Data', 'Cleaned', 'customer_data_wrangled.csv'), index=False)

# Checking basic stats for all numerical fields
insta_cust.agg({'age': ['mean', 'min', 'max']})

insta_cust.agg({'family_members': ['mean', 'min', 'max']})

insta_cust.agg({'income': ['mean', 'min', 'max']})

insta_cust.head()

insta_cust.dtypes

# Importing orders & products data
ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Cleaned', 'ords_prods_merge_mktg.pkl'))

ords_prods.head()

ords_prods = ords_prods.drop(columns=['_merge','busiest_day'])

ords_prods.head()

ords_prods.dtypes

# Merging customer data with ords/prods data
insta_complete = ords_prods.merge(insta_cust, on=['user_id'], indicator = True)

# Checking merge
insta_complete.shape

insta_complete.head()

insta_complete['_merge'].value_counts()

# Testing outer merge
insta_outer = ords_prods.merge(insta_cust, on=['user_id'], how ='outer', indicator = True)

insta_outer['_merge'].value_counts()

# Removing unnecessary column
insta_complete = insta_complete.drop(columns=['_merge'])

insta_complete.head()

# Saving complete merged data
insta_complete.to_pickle(os.path.join(path, '02 Data', 'Cleaned', 'instacart_complete.pkl'))
